% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classificacao_b3}
\alias{classificacao_b3}
\title{Reference data for B3 sector classification}
\format{
data.frame with 6 columns:
\describe{
\item{sector}{Sector of the company}
\item{subsector}{Subsector of the company}
\item{segment}{Segment of the company}
\item{company_name}{Name of the company}
\item{symbol}{Stock ticker symbol}
\item{segment_code}{B3 Code of the segment}
}
}
\source{
\url{https://www.b3.com.br/en_us/products-and-services/trading/equities/equities/reports/industry-classification/}
}
\usage{
classificacao_b3
}
\description{
A dataset containing the sector classification of companies listed on B3, translated to English and standardized for consistency.
}
\details{
This dataset is not updated regularly. It is used to classify brazilian stocks into sectors, subsectors, and segments based on B3's official classification.
For the most current classification, refer to the official B3 website.
}
\keyword{datasets}
