% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{create_earnings_features}
\alias{create_earnings_features}
\title{Create earnings features}
\usage{
create_earnings_features(
  symbols_price_data,
  earnings_calendar,
  eps_history,
  eps_estimates_quarters
)
}
\arguments{
\item{symbols_price_data}{A data frame containing price data for symbols.}

\item{earnings_calendar}{A data frame containing earnings calendar information.}

\item{eps_history}{A data frame containing earnings history information.}

\item{eps_estimates_quarters}{A data frame containing quarterly earnings estimates.}
}
\value{
A data frame with earnings features.
}
\description{
Joins earnings data, consensus estimates, and price data to create features for analysis.
}
