% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_page}
\alias{fake_page}
\title{Fake page content for testing}
\format{
An object of class \code{xml_document} (inherits from \code{xml_node}) of length 2.
}
\usage{
fake_page
}
\description{
A sample HTML content used for testing web scraping functions.
It simulates a webpage structure with stock symbols and company names.
}
\keyword{datasets}
