% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{fix_financial_infinities}
\alias{fix_financial_infinities}
\title{Fix specific problematic columns in existing data}
\usage{
fix_financial_infinities(eps_data)
}
\arguments{
\item{eps_data}{A data frame containing EPS data.}
}
\value{
A data frame with infinite values replaced by specified limits.
}
\description{
Handles infinite values in financial data by applying winsorization.
}
