% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_donchian_channels}
\alias{get_donchian_channels}
\title{Get Donchian Channels (Rolling Min/Max) from Database}
\usage{
get_donchian_channels(db_con, timeframe = "1d", period = 52)
}
\arguments{
\item{db_con}{A DBI connection object.}

\item{timeframe}{Character ("1d" or "1h").}

\item{period}{Integer. Lookback window (default 52).}
}
\value{
A data frame with upper_donchian_P, lower_donchian_P, and mid_donchian_P.
}
\description{
Calculates the rolling High and Low bounds and the Mid-point.
}
