% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_earnings_calendar}
\alias{get_earnings_calendar}
\title{Get the earnings calendar}
\usage{
get_earnings_calendar(db_con)
}
\arguments{
\item{db_con}{A DBI connection object.}
}
\value{
A data frame with earnings calendar information.
}
\description{
Retrieves the earnings calendar with period labels and end dates.
}
