% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_emas}
\alias{get_emas}
\title{Calculate Multiple EMAs for a Timeseries}
\usage{
get_emas(timeseries, periods = c(20, 50, 150, 200))
}
\arguments{
\item{timeseries}{A data frame containing 'symbol', 'close', and 'open_time'.}

\item{periods}{A numeric vector of periods (e.g., 20, 50).}
}
\value{
A grouped_df (or ungrouped df if fixed) with new columns 'ema_P'.
}
\description{
Calculate Multiple EMAs for a Timeseries
}
