% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_eps_data}
\alias{get_eps_data}
\title{Get the earnings history for a specific symbol}
\usage{
get_eps_data(db_con, symbols_price_data)
}
\arguments{
\item{db_con}{A DBI connection object.}

\item{symbols_price_data}{A data frame containing symbols to filter by.}
}
\value{
A data frame with earnings history information.
}
\description{
Retrieves earnings history for a specific symbol.
}
