% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_external_indexes}
\alias{get_external_indexes}
\title{Function to get benchmarks indicators}
\usage{
get_external_indexes(
  db_con,
  timeframe = "1d",
  start_date = NULL,
  end_date = NULL,
  symbols = NULL
)
}
\arguments{
\item{db_con}{A DBI connection object to the database containing the benchmark data.}

\item{timeframe}{A character string specifying the timeframe ("1d" or "1h").}

\item{start_date}{A character string specifying the start date for the data (optional).}

\item{end_date}{A character string specifying the end date for the data (optional).}

\item{symbols}{A character vector of symbols to filter the benchmark data (optional).}
}
\value{
A data frame with the processed benchmark data.
}
\description{
Loads and processes external market indexes data to use as benchmarks.
}
