% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_smas}
\alias{get_smas}
\title{Get the Simple Moving Averages (SMA) for a given set of periods}
\usage{
get_smas(db_con, timeframe = "1d", periods = c(20, 50, 150, 200))
}
\description{
Get the Simple Moving Averages (SMA) for a given set of periods
}
