% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_xgb_indicator}
\alias{get_xgb_indicator}
\title{Get the xgb model as an indicator}
\usage{
get_xgb_indicator(model_conn, model_labels)
}
\arguments{
\item{model_conn}{A DBI connection object to the database containing the xgb_indicator}

\item{model_labels}{A character vector of model labels to filter the xgb_indicator data.}
}
\value{
A data frame with xgb indicator information for the specified model labels.
}
\description{
Retrieves the xgb model predictions as an indicator for the specified model labels.
}
