% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_valid_db_connection}
\alias{is_valid_db_connection}
\title{Check is you have a valid database connection}
\usage{
is_valid_db_connection(db_con)
}
\arguments{
\item{db_con}{A valid DBI database connection object.}
}
\value{
TRUE if the connection is valid, FALSE otherwise.
}
\description{
Check is you have a valid database connection
}
\examples{
\dontrun{
# Create a database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:", read_only = FALSE)
# Check if the connection is valid
is_valid_db_connection(db_con)
# Disconnect from the database
DBI::dbDisconnect(db_con)
}
}
