% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_all_symbols}
\alias{sample_all_symbols}
\title{Sample of stock symbols data}
\format{
An object of class \code{data.frame} with 52 rows and 12 columns.
}
\usage{
sample_all_symbols
}
\description{
A sample dataset containing stock symbols and related information for randomly selected companies.
Its used for testing and demonstration purposes.
}
\keyword{datasets}
