% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_fundamentals}
\alias{sample_fundamentals}
\title{Sample of fundamental data}
\format{
An object of class \code{list} of length 10.
}
\usage{
sample_fundamentals
}
\description{
A sample dataset containing fundamental financial data for "AAPL", "MSFT", "GOOGL".
Its used for testing and demonstration purposes.
}
\keyword{datasets}
