% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_stock_prices}
\alias{sample_stock_prices}
\title{Sample of stock price data}
\format{
An object of class \code{list} of length 2.
}
\usage{
sample_stock_prices
}
\description{
A sample dataset containing stock price data for "AAPL", "MSFT", "GOOGL".
Its used for testing and demonstration purposes.
}
\keyword{datasets}
