% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tables.R
\name{update_benchmarks}
\alias{update_benchmarks}
\title{Create benchmark symbols or add new ones if they don't exist}
\usage{
update_benchmarks(db_con, new_benchmarks = NULL)
}
\value{
Nothing. The function updates the 'benchmark_symbols' table in the database.
}
\description{
This function creates a predefined set of benchmark symbols with their associated information.
If the benchmark symbols already exist in the database, it checks for any new symbols and adds them.
Despite having the same structure as 'all_symbols', the 'benchmark_symbols' are not expecting data changes, therefore, no updates are usually needed.
}
\examples{
\dontrun{
# define the path and name of the database file
db_name <- "test_stock_db"
db_path <- file.path(tempdir(), db_name)
# create the database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = db_path, read_only = FALSE)
# Create or update benchmark symbols
update_benchmarks(db_con = db_con)
DBI::dbDisconnect(db_con)
}
}
