% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dividends.R
\name{update_dividends}
\alias{update_dividends}
\title{Update the dividends table in the database}
\usage{
update_dividends(db_con, symbols_to_get = NULL, start_date = "2014-01-01")
}
\arguments{
\item{db_con}{A valid DBI database connection object.}

\item{symbols_to_get}{A character vector of stock symbols to update.
If NULL, all symbols from \code{load_all_symbols(db_con)$all_symbol_data} are used.}

\item{start_date}{A character string "YYYY-MM-DD" (default: "2014-01-01").}
}
\description{
This function updates the 'dividends' table
in a DuckDB database with the latest dividends paid for the selected symbols.
}
\details{
It uses the \code{invest.data} package to fetch new or missing data, while
ensuring no duplication and maintaining data consistency.
}
\examples{
\dontrun{
# 1 Creating the first table
## define the path and name of the database file
db_name <- "test_stock_db"
db_path <- file.path(tempdir(), db_name)
## create the database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = db_path, read_only = FALSE)
## Get the important variables
symbols_to_get <- c("A8G.AX", "AAPL", "MSFT")
start_date <- as.character(Sys.Date() - days(600))
interval <- "1d"
## Create the daily prices table in this new database
update_dividends(
db_con,
symbols_to_get,
start_date
)
## Check the contents of the dividends table
loaded_data <- DBI::dbGetQuery(db_con, "SELECT * FROM dividends")
# 2 Updating the existing table with new data
## Fake the data to be old
fake_old_date <- loaded_data \%>\% filter(date < as.character(Sys.Date() - 200))
## write back the fake old data to the database
DBI::dbWriteTable(db_con, "dividends", fake_old_date, overwrite = TRUE)
## Check the contents of the dividends table
symbols_before <- dbGetQuery(db_con,"SELECT symbol, MIN(date) as first_date, MAX(date) as last_date FROM dividends GROUP BY symbol")
## Now run the update function again to get the missing recent data
symbols_to_get <- c("A8G.AX", "AAPL", "MSFT", "GOOGL")
update_dividends(db_con, symbols_to_get, start_date = "2020-01-01")
## Check the table contents after updating
symbols_after <- dbGetQuery(db_con,"SELECT symbol, MIN(date) as first_date, MAX(date) as last_date FROM dividends GROUP BY symbol")

DBI::dbDisconnect(db_con)
}
}
