% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prices.R
\name{update_stock_prices}
\alias{update_stock_prices}
\title{Update the daily or hourly prices table in the database}
\usage{
update_stock_prices(
  db_con,
  symbols_to_get = NULL,
  start_date = "2014-01-01",
  interval = "1d"
)
}
\arguments{
\item{db_con}{A valid DBI database connection object.}

\item{symbols_to_get}{A character vector of stock symbols to update.
If NULL, all symbols from \code{load_all_symbols(db_con)$all_symbol_data} are used.}

\item{start_date}{A character string "YYYY-MM-DD" (default: "2014-01-01").}

\item{interval}{A character string, either "1d" or "1h" (default: "1d").}
}
\description{
This function updates the 'daily_prices' or 'hourly_prices' table
in a DuckDB database with the latest stock prices for the selected symbols.
}
\details{
It uses the \code{invest.data} package to fetch new or missing data, while
ensuring no duplication and maintaining data consistency.
}
\examples{
\dontrun{
# define the path and name of the database file
db_name <- "test_stock_db"
db_path <- file.path(tempdir(), db_name)
# create the database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = db_path, read_only = FALSE)
# Get the important variables
symbols_to_get <- c("A8G.AX", "AAPL", "MSFT")
start_date <- as.character(Sys.Date() - days(200))
interval <- "1d"
# Create the daily prices table in this new database
update_stock_prices(
db_con,
symbols_to_get,
start_date,
interval
)
loaded_data <- DBI::dbGetQuery(db_con, "SELECT * FROM daily_prices")

DBI::dbDisconnect(db_con)
}
}
