% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tables.R
\name{update_symbols_table}
\alias{update_symbols_table}
\title{Update the table with all symbols basic information}
\usage{
update_symbols_table(indexes = NULL, db_con)
}
\arguments{
\item{indexes}{A character vector specifying which indexes to filter by. Options include "ASX", "B3", "SP500", and "Benchmark". If NULL, returns all symbols.}

\item{db_con}{A DBI database connection object.}
}
\value{
Nothing. The function updates the 'all_symbols' table in the database, informing the user of the update status.
}
\description{
This function updates the 'all_symbols' table in the database with the latest stock symbols and their basic information from various indexes.
}
\details{
It fetches data from the ASX, B3, and S&P 500 indexes, combines them, and updates the database table accordingly.
}
\examples{
\dontrun{
# Establish a database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = "path_to_your_db.duckdb", read_only = FALSE)
# Update the symbols table
update_symbols_table(indexes = c("ASX", "B3"), db_con = db_con)
DBI::dbDisconnect(db_con)
}
}
