% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_functions.R
\name{get_emas}
\alias{get_emas}
\title{Get the Exponential Moving Averages (EMA) for a given set of periods}
\usage{
get_emas(timeseries, periods = c(20, 50, 150, 200))
}
\description{
Get the Exponential Moving Averages (EMA) for a given set of periods
}
