% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{load_symbols}
\alias{load_symbols}
\title{Function to get all stock symbols from the DuckDB database}
\usage{
load_symbols(db_con)
}
\arguments{
\item{db_con}{A valid DBI connection object to the DuckDB database.}
}
\value{
A list containing data frames for all symbols and filtered symbols by index.
}
\description{
This function connects to a DuckDB database, retrieves all stock symbols,
and filters them based on their respective indices (ASX, B3, SP500, Benchmark).
}
\examples{
\dontrun{
db_file <- "/mnt/nas_nuvens/stock_data/stock_data - Copia.duckdb"
# create a connection to the database
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = db_file, read_only = TRUE)
symbols <- load_symbols(db_con)
head(symbols$all_symbols_au)
head(symbols$all_symbols_br)
head(symbols$all_symbols_sp500)
head(symbols$all_benchmarks)
dbDisconnect(db_con)
}

}
